+++ src/Clusterer.h
@@ -39,7 +39,7 @@ public:
   bool AnalyzeHits(double srsTimestamp, uint8_t fecId, uint8_t vmmId,
                    uint16_t chNo, uint16_t bcid, uint16_t tdc, uint16_t adc,
                    bool overThresholdFlag, double chipTime, uint8_t geoId = 0,
-                   int event_counter=0);
+                   double event_counter=0);
 
   // Analyzing and storing the clusters in plane 0 and 1
   void AnalyzeClustersPlane(std::pair<uint8_t, uint8_t> dp);
